+++ app/models/github/put/Commiter.scala
@@ -1,4 +1,4 @@
-package models.github.create
+package models.github.put
 
 import models.github.RepoFileItem
 import play.api.libs.json.{Json, OFormat}
